refactor(api): move temporary storage to src/shared
